# public/openapi.yaml
openapi: 3.1.0
info:
  title: DBackup API
  version: 3.2.0
  description: |
    REST API for DBackup - a self-hosted database backup automation platform with encryption, compression, and smart retention.

    ## Authentication

    DBackup supports two authentication methods:

    - **Session Authentication** - Used automatically when logged in via the web UI. Session cookies are sent with each request.
    - **API Key Authentication** - For scripts, CI/CD pipelines, and external integrations. Create an API key under **Access Management → API Keys**.

    > API keys do not inherit SuperAdmin privileges. Only explicitly assigned permissions are available.

    ## Rate Limits

    | Request Type | Default Limit |
    |---|---|
    | Authentication (login, etc.) | 5/min per IP |
    | GET requests | 100/min per IP |
    | POST / PUT / DELETE | 20/min per IP |

    Rate limits are configurable in Settings → Rate Limits.
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
...
servers:
  - url: "{protocol}://{host}/api"
    description: DBackup Instance
    variables:
      protocol:
        default: https
        enum: [http, https]
      host:
        default: your-dbackup-instance.com
security:
  - BearerAuth: []
  - SessionAuth: []
...
paths:
...
  /adapters/{id}:
    put:
      tags: [Adapters]
      summary: Update Adapter
      operationId: updateAdapter
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdapterRequest"
      responses:
        "200":
          description: Adapter updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdapterConfig"
        "409":
          description: Another adapter with this name and type already exists
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Adapters]
      summary: Delete Adapter
      description: Deletes an adapter configuration. Returns error 400 if the adapter is still referenced by active jobs.
      operationId: deleteAdapter
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Adapter deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Adapter is still in use by jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /adapters/{id}/clone:
    post:
      tags: [Adapters]
      summary: Clone Adapter
      description: |
        Creates a copy of the adapter configuration with the suffix `(Copy)` appended to
        the name. The same write permission as the adapter type is required:
        - `database` - `sources:write`
        - `storage` - `destinations:write`
        - `notification` - `notifications:write`
      operationId: cloneAdapter
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "201":
          description: Cloned adapter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdapterConfig"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API Key authentication. Create an API key under **Access Management → API Keys**.

        Format: `dbackup_xxxxxxxx`
    SessionAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Session cookie (browser login)
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource ID
  responses:
    Unauthorized:
      description: Missing, invalid, disabled, or expired credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Valid credentials but insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Human-readable error message"
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
...
    AdapterConfig:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: "Production MySQL"
        type:
          type: string
          enum: [database, storage, notification]
        adapterId:
          type: string
          description: Adapter type identifier
          example: mysql
        config:
          type: object
          description: Adapter-specific configuration (decrypted)
        metadata:
          type: [object, "null"]
          description: "Auto-detected metadata (version, status, etc.)"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        storageRole:
          type: string
          enum: [DESTINATION, SOURCE]
          default: DESTINATION
          description: >-
            Storage adapters only. A config either receives backups (DESTINATION) or
            provides folders to back up (SOURCE) - never both, because a destination writes
            job and chain folders into its configured path while a source reads folders out
            of it. Ignored for database and notification adapters.
...
    UpdateAdapterRequest:
      type: object
      properties:
        name:
          type: string
        config:
          type: object
          description: Updated configuration (merged with existing)
        storageRole:
          type: string
          enum: [DESTINATION, SOURCE]
          description: >-
            Storage adapters only. Rejected with 400 while a job still uses the adapter in
            its current role, since the change would remove it from that job's picker.

    # ─── Storage ─────────────────────────────────────────────────
...